strip down SpotifyMediaClass and improve windows behaviour
