+++ js/SpotifyMediaClass/emitter.ts
