+++ js/components/Explorer/styles.js
@@ -14,10 +14,8 @@ const ExplorerWindowStyle = {
     flexDirection: "column",
     height: "100%",
     width: "100%",
-    border: "1px solid rgba(0, 0, 0, 1)",
-    borderRadius: "2px",
     "box-shadow":
-      "0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19)"
+      "0 4px 8px 0 rgba(0, 0, 0, 0.4), 0 6px 20px 0 rgba(0, 0, 0, 0.30)"
   },
 
   explorerToolbar: {
