+++ js/reducers/windows.ts
@@ -1,15 +1,24 @@
 import _ from "lodash";
-import { OPEN_EXPLORER } from "./explorer";
+import {
+  OPEN_EXPLORER,
+  CLOSE_EXPLORER,
+  UPDATE_POSITION,
+  UPDATE_SIZE
+} from "./explorer";
 import { OPEN_IMAGE, CLOSE_IMAGE } from "../actions/images";
 
-export enum WindowType {
+export enum WINDOW_TYPE {
   Explorer,
   Image
 }
 
 export interface Window {
   id: string;
-  type: WindowType;
+  type: WINDOW_TYPE;
+  /*   x: number;
+  y: number;
+  height: number;
+  width: number; */
 }
 
 export interface WindowsState {
@@ -24,12 +33,32 @@ const windows = (state: WindowsState = initialState, action: any) => {
   switch (action.type) {
     case OPEN_EXPLORER: {
       const { windows } = state;
-      windows.push({ id: action.id, type: WindowType.Explorer });
+      windows.push({
+        id: action.payload.id,
+        type: WINDOW_TYPE.Explorer
+        /*  x: action.payload.x,
+        y: action.payload.y,
+        height: 500,
+        width: 400 */
+      });
       return { windows };
+    }
+    case CLOSE_EXPLORER: {
+      const { windows } = state;
+      return {
+        windows: windows.filter(window => window.id !== action.payload.id)
+      };
     }
     case OPEN_IMAGE: {
       const windows = state.windows;
-      windows.push({ id: action.id, type: WindowType.Image });
+      windows.push({
+        id: action.id,
+        type: WINDOW_TYPE.Image
+        /*  x: action.payload.x,
+        y: action.payload.y,
+        height: 200,
+        width: 200 */
+      });
       return { windows };
     }
     case CLOSE_IMAGE: {
@@ -47,6 +76,26 @@ const windows = (state: WindowsState = initialState, action: any) => {
       windows.push(windowToMove);
       return { windows };
     }
+    /* case UPDATE_POSITION:
+      return {
+        ...state,
+        byId: {
+          ...state.byId,
+          [action.payload.id]: {
+            ...state.byId[action.payload.id]
+          }
+        }
+      };
+    case UPDATE_SIZE:
+      return {
+        ...state,
+        byId: {
+          ...state.byId,
+          [action.id]: {
+            ...state.byId[action.id]
+          }
+        }
+      }; */
     default:
       return state;
   }
