--- js/spotifyEvents.ts
@@ -1,4 +1,0 @@
-import EventEmitter from "events";
-const eventEmitter = new EventEmitter();
-
-export default eventEmitter;
